Fuzzing campaign: 23 engine fixes, 5 new fuzz targets, recovery hardening
Whole-engine libFuzzer campaign (12 waves, ~120 CPU-hours, UBSan+ASan
builds); full write-up in docs/fuzzing_campaign_202609.md.

Engine bugs found & fixed (each reproduced from a saved unit, re-verified,
2410/2410 ctest green after every batch):
- bytecode VM: AND/OR short-circuit now returns AST-normalized BOOLEAN;
  truthiness-collapse folds removed from kValue (value/type contexts)
- rewrite: double_negation / x+x / OR-distribution INT64-overflow and
  throw-placement guards; NaN-soundness cluster (not_comparison now
  schema-aware via thread-local registration at bytecode/scan/optimizer)
- IN comparison-error propagation; IS [NOT] DISTINCT FROM result type;
  Value::Truthy(0.0)
- TRUNC/ROUND INT64_MIN digit negate; REPEAT overflow + empty-string loop
- interval: 5 double->int64 UB sites guarded to exact [-2^63, 2^63)
- recovery: seed next txn id from WAL max (old commit records no longer
  vouch for new transactions); PageReplay skips ops a CLR already closed
- frontend: throwing IntervalExpr/make_interval/CAST INTERVAL/proto-text
  paths converted to Status (Try family)

Fuzz infrastructure:
- project-wide -fsanitize=fuzzer-no-link,undefined,bounds in fuzz builds
  (SQL-target coverage 1 -> 22k PCs, feedback-guided corpora)
- new targets: pax_page, date_interval, cast, function, data_chunk (25)
- crash/recovery model ops in leaf-page / B+tree / LSM harnesses
- fuzzer-nightly runs all 12 in-process targets with clang pin

Also folds in pending work: window function executor, cascades rule
migration fixes, and the docs/opt optimizer explainer series.

+++ common/serdes.cpp
@@ -49,8 +49,9 @@ size_t SerializeU64(char* pos, uint64_t value) {
 }
 
 size_t DeserializeU16(const char* pos, uint16_t* out) {
-  *out = (static_cast<uint16_t>(static_cast<unsigned char>(pos[0])) << 8U) |
-         static_cast<uint16_t>(static_cast<unsigned char>(pos[1]));
+  *out = static_cast<uint16_t>(
+      (static_cast<uint16_t>(static_cast<unsigned char>(pos[0])) << 8U) |
+      static_cast<uint16_t>(static_cast<unsigned char>(pos[1])));
   return sizeof(*out);
 }
